Clear the decks for battle.
Updated the deps that didn't require thought to update.
Added some lint rules and fixed all the `cargo check` violations.

Replaced all the `.unwrap()` usage in the cli crate with `.expect()`
and specific error messages. There are 199 usages of unwrap in the
library code that should not exist IMO because it's a library and should
have more well-defined error handling. I'll do those next.